Modularized static_site_builder
